requirements.txt
# Core
python-dotenv==1.2.1
pandas==2.2.2

# Kafka
kafka-python==2.3.0

# Async HTTP (producer)
aiohttp==3.13.2

# Airflow
apache-airflow==3.1.3

# Utils
requests==2.32.5
sqlite3==3.45.3